Update 4.1.2 Fixed bug caused the extension to fail to load
With the current public version 4.1.0 we introduced an issue where the extension fails to load entirely on new installations of the addin.

The main reason for this was within the function "addCPSToJSLanguage" which failed to execute properly in cases where the system had no file associations defined yet.
JSON.parse did fail to execute and the extension did not work at all.

In addition, fixed multiple minor issues at the same time.